# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Carpoolink API
  version: 0.1.0
  description: 사용자 식별이 필요한 API는 `x-user-id` 헤더가 필요합니다.
servers:
  - url: http://localhost:4000
  - url: http://localhost:4002
paths:
  /questions/recommendations:
    port: 4000
    post:
      summary: 질문 추천 생성 프록시
      description: 로그인 사용자의 멘토링 세션 접근 권한을 확인하고, DB의 멘티/멘토 정보를 조합해 question-service의 AI 질문 추천 API를 호출합니다. `x-user-id` 헤더가 필요합니다.
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
          description: 현재 사용자 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sessionId
                - context
              properties:
                sessionId:
                  oneOf:
                    - type: integer
                    - type: string
                  example: 1
                  description: 멘토링 세션 ID
                topic:
                  type: string
                  example: React 상태 관리
                  description: 생략하면 멘토링 제목을 사용합니다.
                context:
                  type: string
                  example: useState와 useReducer의 차이를 학습 중이며, 어떤 상황에서 useReducer를 써야 하는지 궁금해하고 있습니다.
                count:
                  type: integer
                  minimum: 1
                  maximum: 5
                  default: 3
                  example: 3
      responses:
        '200':
          description: 추천 질문 생성 성공
          content:
            application/json:
              schema:
                type: object
                required:
                  - questions
                properties:
                  questions:
                    type: array
                    items:
                      type: object
                      required:
                        - content
                        - category
                        - reason
                      properties:
                        content:
                          type: string
                          example: useState 대신 useReducer를 사용하는 기준은 무엇인가요?
                        category:
                          type: string
                          enum:
                            - concept
                            - reasoning
                            - application
                            - comparison
                            - follow_up
                          example: comparison
                        reason:
                          type: string
                          example: 멘티 수준과 멘토 전문 분야에 맞는 후속 질문입니다.
        '400':
          description: 입력값 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: QUESTION_RECOMMENDATION_INVALID_INPUT
                  message:
                    type: string
                    example: context is required
        '404':
          description: 세션 없음 또는 접근 불가
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: QUESTION_RECOMMENDATION_SESSION_NOT_FOUND
                  message:
                    type: string
                    example: mentoring session not found or access denied
        '502':
          description: question-service 또는 AI 호출 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: QUESTION_RECOMMENDATION_GENERATION_FAILED
                  message:
                    type: string
                    example: failed to generate question recommendations
  /api/questions/recommendations:
    port: 4003
    post:
      summary: 질문 추천 생성
      description: 멘토링 맥락, 멘티 정보, 멘토 정보를 기반으로 AI가 추천 질문을 생성합니다. mock 또는 rule 기반 fallback은 사용하지 않습니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - context
                - mentee
                - mentor
              properties:
                sessionId:
                  type: integer
                  nullable: true
                  example: 1
                topic:
                  type: string
                  example: React 상태 관리
                context:
                  type: string
                  example: useState와 useReducer의 차이를 학습 중이며, 어떤 상황에서 useReducer를 써야 하는지 궁금해하고 있습니다.
                count:
                  type: integer
                  minimum: 1
                  maximum: 5
                  default: 3
                  example: 3
                mentee:
                  type: object
                  required:
                    - id
                  properties:
                    id:
                      oneOf:
                        - type: integer
                        - type: string
                      example: 10
                    name:
                      type: string
                      example: 김민지
                    level:
                      type: string
                      example: beginner
                    interests:
                      type: array
                      items:
                        type: string
                      example:
                        - frontend
                        - React
                    goals:
                      type: array
                      items:
                        type: string
                      example:
                        - React 상태 관리 이해
                        - 프로젝트에 적용
                mentor:
                  type: object
                  required:
                    - id
                  properties:
                    id:
                      oneOf:
                        - type: integer
                        - type: string
                      example: 3
                    name:
                      type: string
                      example: 이준호
                    expertise:
                      type: array
                      items:
                        type: string
                      example:
                        - frontend
                        - React
                        - architecture
                    role:
                      type: string
                      example: senior frontend engineer
      responses:
        '200':
          description: 추천 질문 생성 성공
          content:
            application/json:
              schema:
                type: object
                required:
                  - service
                  - questions
                properties:
                  service:
                    type: string
                    example: question-service
                  questions:
                    type: array
                    items:
                      type: object
                      required:
                        - content
                        - category
                        - reason
                      properties:
                        content:
                          type: string
                          example: useState 대신 useReducer를 사용하는 기준은 무엇인가요?
                        category:
                          type: string
                          enum:
                            - concept
                            - reasoning
                            - application
                            - comparison
                            - follow_up
                          example: comparison
                        reason:
                          type: string
                          example: 멘티가 beginner 수준이고 React 상태 관리 개념을 학습 중이기 때문입니다.
        '400':
          description: 입력값 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: QUESTION_RECOMMENDATION_INVALID_INPUT
                  message:
                    type: string
                    example: context is required
        '502':
          description: AI 호출 또는 응답 파싱 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: QUESTION_RECOMMENDATION_GENERATION_FAILED
                  message:
                    type: string
                    example: failed to generate question recommendations
...